pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ms-ai-classifier"
version = "0.1.0"
description = "AI classifier microservice for Mi Cartera bank statement ingestion"
requires-python = ">=3.12"
dependencies = [
  "aio-pika==9.5.4",
  "fastapi==0.115.12",
  "httpx==0.28.1",
  "openai==1.82.0",
  "pdfplumber==0.11.6",
  "pydantic-settings==2.9.1",
  "pymupdf==1.26.0",
  "python-multipart==0.0.20",
  "sqlalchemy==2.0.40",
  "psycopg[binary]==3.2.9",
  "uvicorn[standard]==0.34.2"
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.5"
]

[tool.setuptools.packages.find]
include = ["app*"]
